Move some mechanical visitors to dedicated folder
